pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "test-time-hinting"
version = "0.1.0"
description = "Hint optimization pipeline for Test-Time Hinting (Proposer / Checker / Experimenter)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "openai>=1.0.0",
    "google-genai>=1.0.0",
    "tqdm>=4.65.0",
    "anthropic>=0.39.0",
]

[project.scripts]
tth = "tth.main:main"

[tool.setuptools.packages.find]
where = ["src"]
